[8098] Support uint32 spell ids in code.
* Propertly work with uint32 spell ids in player action bar
* Fix in same time bug with not save equipment set button with id==0
* Merge misc field in character_action and playercreateinfo_action to action field as 3 byte
* Propertly load uint32 spell ids from character_spell
* Fixed types for some pet/creature related structure for spell id storing.

(cherry picked from commit 1906637785583a98c8e0be416dec4b6af2000829)

+++ src/game/SpellMgr.cpp
@@ -1961,9 +1961,9 @@ void SpellMgr::LoadSpellPetAuras()
 
         bar.step();
 
-        uint16 spell = fields[0].GetUInt16();
-        uint16 pet = fields[1].GetUInt16();
-        uint16 aura = fields[2].GetUInt16();
+        uint32 spell = fields[0].GetUInt32();
+        uint32 pet = fields[1].GetUInt32();
+        uint32 aura = fields[2].GetUInt32();
 
         SpellPetAuraMap::iterator itr = mSpellPetAuraMap.find(spell);
         if(itr != mSpellPetAuraMap.end())
